recipe_transformation/requirements.txt
pandas~=2.2.2
jinja2~=3.1.4
pydantic~=2.5.3
python-multipart~=0.0.20
uvicorn~=8.1.7
tortoise-orm~=0.24.2